refactor: SaveReviewWithLatLng 클래스 data.firebase 패키지로 이동
